Move server-side scripts into root folder
